Refactored common types
