+++ docker-compose.yml
@@ -14,6 +14,7 @@ services:
       NODE_ENV: development
     volumes:
       - ./frontend:/opt/app
+      - ./types:/opt/types
       - /opt/app/node_modules
     depends_on:
       - backend
@@ -30,6 +31,7 @@ services:
       REDIS_URL: redis://redis:6379/
     volumes:
       - ./backend:/opt/app
+      - ./types:/opt/types
       - /opt/app/node_modules
     depends_on:
       - redis
